+++ spec/rubocop/cop/rspec/shared_examples_spec.rb
@@ -164,4 +164,111 @@ RSpec.describe RuboCop::Cop::RSpec::SharedExamples do
       end
     end
   end
+
+  context 'with preferred methods' do
+    shared_examples 'preferred method unset' do |config_key, methods|
+      context "when #{config_key} is unset" do
+        let(:config_val) { nil }
+
+        methods.each do |method|
+          it "does not register an offense for `#{method}`" do
+            expect_no_offenses(<<~RUBY)
+              #{method} 'desc'
+            RUBY
+          end
+        end
+
+        it 'does not register an offense for another method' do
+          expect_no_offenses(<<~RUBY)
+            describe
+          RUBY
+        end
+      end
+    end
+
+    shared_examples 'preferred method set' do |config_key, methods, preferred|
+      context "when #{config_key} is set" do
+        let(:config_val) { preferred }
+
+        it "does not register an offense for `#{preferred}`" do
+          expect_no_offenses(<<~RUBY)
+            #{preferred} 'desc'
+          RUBY
+        end
+
+        (methods - [preferred]).each do |method|
+          it "registers an offense and corrects for #{method}" do
+            expect_offense(<<~RUBY, method: method, preferred: preferred)
+              %{method} 'desc'
+              ^{method} Prefer `%{preferred}` over `%{method}`.
+            RUBY
+
+            expect_correction(<<~RUBY)
+              #{preferred} 'desc'
+            RUBY
+          end
+        end
+
+        it 'does not register an offense for another method' do
+          expect_no_offenses(<<~RUBY)
+            describe
+          RUBY
+        end
+
+        it 'registers offenses and corrects for both issues ' \
+           'when `EnforcedStyle` also offends' do
+          method = (methods - [preferred]).first
+          expect_offense(<<~RUBY, method: method, preferred: preferred)
+            %{method} :desc
+            ^{method} Prefer `%{preferred}` over `%{method}`.
+            _{method} ^^^^^ Prefer 'desc' over `:desc` to titleize shared examples.
+          RUBY
+
+          expect_correction(<<~RUBY)
+            #{preferred} 'desc'
+          RUBY
+        end
+      end
+    end
+
+    shared_examples 'preferred method' do |config_key, methods, preferred|
+      context "with #{config_key}" do
+        let(:cop_config) do
+          { 'EnforcedStyle' => 'string', config_key => config_val }
+        end
+
+        it_behaves_like 'preferred method unset', config_key, methods
+        it_behaves_like 'preferred method set', config_key, methods, preferred
+      end
+    end
+
+    before do
+      # Set up custom aliases
+      other_cops.dig('RSpec', 'Language', 'SharedGroups', 'Examples')
+        .push('shared_scenarios')
+      other_cops.dig('RSpec', 'Language', 'SharedGroups', 'Context')
+        .push('shared_state')
+      other_cops.dig('RSpec', 'Language', 'Includes', 'Examples')
+        .push('include_scenarios')
+      other_cops.dig('RSpec', 'Language', 'Includes', 'Context')
+        .push('include_state')
+    end
+
+    it_behaves_like 'preferred method', 'PreferredExamplesMethod',
+                    %w[shared_examples shared_examples_for shared_scenarios],
+                    'shared_scenarios'
+
+    it_behaves_like 'preferred method', 'PreferredContextMethod',
+                    %w[shared_context shared_state],
+                    'shared_state'
+
+    it_behaves_like 'preferred method', 'PreferredIncludeExamplesMethod',
+                    %w[it_behaves_like it_should_behave_like include_examples
+                       include_scenarios],
+                    'include_scenarios'
+
+    it_behaves_like 'preferred method', 'PreferredIncludeContextMethod',
+                    %w[include_context include_state],
+                    'include_state'
+  end
 end